requirements.txt
streamlit>=1.35.0
pymongo>=4.7.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
Pillow>=10.4.0
pdfplumber>=0.11.0
python-docx>=1.1.0
plotly>=5.22.0
pandas>=2.2.0
dnspython>=2.6.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
numpy>=1.26.0
streamlit-extras>=0.4.0
bcrypt>=4.2.0
requests>=2.32.0
jinja2>=3.1.0
cryptography>=44.0.0
markdown>=3.6.0
